+++ lib/td_lm_web/controllers/graph_controller.ex
@@ -16,21 +16,22 @@ defmodule TdLmWeb.GraphController do
 
   def graph(conn, %{"resource_id" => id, "type" => type}) do
     claims = conn.assigns[:current_resource]
+    locale = conn.assigns[:locale]
 
     with {:can, true} <- {:can, can?(claims, show(%{resource_type: type, resource_id: id}))},
          %{nodes: nodes, edges: edges} <-
            Resources.graph(claims, id, type, types: ["business_concept"]) do
-      nodes = enrich_nodes(nodes)
+      nodes = enrich_nodes(nodes, lang: locale)
       render(conn, "show.json", graph: %{nodes: nodes, edges: edges})
     end
   end
 
-  defp enrich_nodes(nodes) do
-    Enum.map(nodes, &enrich_node/1)
+  defp enrich_nodes(nodes, opts) do
+    Enum.map(nodes, &enrich_node(&1, opts))
   end
 
-  defp enrich_node(%{resource_id: id, resource_type: "business_concept"} = node) do
-    {:ok, c} = ConceptCache.get(id)
+  defp enrich_node(%{resource_id: id, resource_type: "business_concept"} = node, opts) do
+    {:ok, c} = ConceptCache.get(id, opts)
 
     case c do
       nil ->
@@ -43,7 +44,7 @@ defmodule TdLmWeb.GraphController do
     end
   end
 
-  defp enrich_node(%{resource_id: id, resource_type: "ingest"} = node) do
+  defp enrich_node(%{resource_id: id, resource_type: "ingest"} = node, _opts) do
     {:ok, i} = IngestCache.get(id)
 
     node
@@ -51,10 +52,10 @@ defmodule TdLmWeb.GraphController do
     |> Map.put(:version_id, Map.get(i, :ingest_version_id))
   end
 
-  defp enrich_node(%{resource_id: id, resource_type: "data_structure"} = node) do
+  defp enrich_node(%{resource_id: id, resource_type: "data_structure"} = node, _opts) do
     {:ok, d} = StructureCache.get(id)
     Map.put(node, :name, Map.get(d, :name))
   end
 
-  defp enrich_node(node), do: node
+  defp enrich_node(node, _opts), do: node
 end